assignment-1-784915/code/requirements.txt
cassandra-driver
numpy
pandas
argparse
matplotlib
tqdm
